fix(serialization): normalize function stringification
A non-default option for serialization allows for the automatically
stringifying javascript functions. Function.prototype.toString has
changed between versions of node, so we need to normalize so as to
not potentially break compatibility with earlier expectations. This
should have marginal performance effects only in the case where
users are actually using this functionality.

NODE-1499