feat(BSON): simplify and flatten module exports
For some time we have been stuck with BSON being an exported
function that requires instantiation, and then has a number of
attached properties all for legacy reasons. With the impending
major version bump, we can remove this requirement and export
everything at the top level of the module, greatly simplifying
user consumption, bundling, and maintenance.