Fix getTimestamp for times beyond 2038
Avoid treating timestamps with the most significant bit set as negative. The mongodb documentation describe the timetsamp portion of objectids as 'seconds since the unix epoch', so this seems to be the correct behaviour, although I cannot find a formal specification for it.however this is technically a breaking change if people were relying on timestamp values before 1970.